changing role done in back and front

+++ GamePortal/Kbalan.TouchType.Client/TTGClient/src/app/user/usermanager/usermanager.component.ts
@@ -3,6 +3,10 @@ import { Router } from '@angular/router';
 import { UserService } from '../services/user.service';
 import { ToastrService } from 'ngx-toastr';
 import { TableUserDto } from '../models/TableUserDto';
+import { LoginService } from 'src/app/core/services/login.service';
+import { UserDto } from 'src/app/core/models/UserDto';
+import { Observable } from 'rxjs';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-usermanager',
@@ -14,25 +18,17 @@ export class UsermanagerComponent implements OnInit {
   allIncomeUsers: TableUserDto[];
   selectedAll: any;
   allUsers =  new Array();
+  user$: Observable<UserDto>;
   @ViewChild('readOnlyTemplate', {static: false}) readOnlyTemplate: TemplateRef<any>;
-  constructor(private router: Router, private userservice: UserService, private toastr: ToastrService) { }
+  constructor(
+    private router: Router,
+    private userservice: UserService,
+    private toastr: ToastrService,
+    private loginService: LoginService) { }
 
   ngOnInit(): void {
-    this.userservice.getAllUsers().subscribe(
-      res => {
-        this.allIncomeUsers = res;
-        for (const num of this.allIncomeUsers) {
-          this.allUsers.push(  {
-            id: num.Id,
-            username: num.UserName,
-            regdate: num.RegistrationDate,
-            logdate: num.LastLoginDate,
-            blocked: num.IsBlocked,
-            role: num.UserRole,
-            selected: false,
-           });
-       }
-      });
+    this.initUsers();
+    this.user$ = this.loginService.LoggedOn$;
   }
 
   loadTemplate(allIncomeUsers) {
@@ -63,16 +59,39 @@ export class UsermanagerComponent implements OnInit {
   }
 
   onBlock(){
+    // tslint:disable-next-line: prefer-for-of
     for (let i = 0; i < this.allUsers.length; i++) {
       if (this.allUsers[i].selected === true)
       {
-        this.allUsers[i].blocked = true;
-        this.userservice.block(this.allUsers[i].id).subscribe();
+        if (this.allUsers[i].id !== this.loginService.LoggedOn.sub)
+        {
+          this.allUsers[i].blocked = true;
+          this.userservice.block(this.allUsers[i].id).subscribe();
+        }
+        else{
+          Swal.fire({
+            title: 'Do you want to block yourself?',
+            text: 'You will not be able to unblock you!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, block me!'
+          }).then((willDelete) => {
+          if (willDelete.isConfirmed) {
+            this.allUsers[i].blocked = true;
+            this.userservice.block(this.allUsers[i].id).subscribe();
+            this.loginService.logout();
+          }
+        });
+        }
       }
     }
   }
 
+
   onUnBlock(){
+    // tslint:disable-next-line: prefer-for-of
     for (let i = 0; i < this.allUsers.length; i++) {
       if (this.allUsers[i].selected === true)
       {
@@ -83,12 +102,59 @@ export class UsermanagerComponent implements OnInit {
   }
 
   onDelete(){
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to revert this!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((willDelete) => {
+    if (willDelete) {
     for (let i = 0; i < this.allUsers.length; i++) {
       if (this.allUsers[i].selected === true)
       {
-        this.userservice.delete(this.allUsers[i].id).subscribe();
-        this.allUsers.splice(i, i);
+        if (this.allUsers[i].id !== this.loginService.LoggedOn.sub)
+        {
+          this.userservice.delete(this.allUsers[i].id).subscribe();
+        }else
+        {
+          Swal.fire({
+            title: 'Are you sure you want to delete yourself?',
+            text: 'You will not be able to revert this!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((deleteYourSelf) => {
+          if (deleteYourSelf.isConfirmed) {
+            this.userservice.delete(this.allUsers[i].id).subscribe();
+            this.loginService.logout();
+          }
+          });
+        }
+        this.allUsers.splice(i, 1);
         i--;
+      }
+    }
+  }
+  });
+  }
+
+  onRoleChange(){
+    // tslint:disable-next-line: prefer-for-of
+    for (let i = 0; i < this.allUsers.length; i++) {
+      if (this.allUsers[i].selected === true)
+      {
+        if (this.allUsers[i].role === 'user'){
+          this.allUsers[i].role  = 'administrator';
+          this.userservice.mkRoleAdmin(this.allUsers[i].id).subscribe();
+        } else{
+          this.allUsers[i].role  = 'user';
+          this.userservice.mkRoleUser(this.allUsers[i].id).subscribe();
+        }
       }
     }
   }
